+++ client/src/components/active-911-calls/index.tsx
@@ -75,6 +75,7 @@ const Active911Calls: React.FC<Props> = ({
               <th scope="col">{window.lang.dispatch.caller_location}</th>
               <th scope="col">{window.lang.dispatch.call_desc}</th>
               <th scope="col">{window.lang.dispatch.status}</th>
+              <th scope="col">{window.lang.citizen.medical.type2}</th>
               <th scope="col">{window.lang.dispatch.assigned_unit}</th>
               <th scope="col">{window.lang.global.actions}</th>
             </tr>
@@ -88,6 +89,7 @@ const Active911Calls: React.FC<Props> = ({
                   <td>{call.location}</td>
                   <td>{call.description}</td>
                   <td>{call.status}</td>
+                  <td>{call.type === "1" ? window.lang.dispatch.citizen_call : call.type}</td>
                   <td>
                     {call.assigned_unit.map((c, i) => {
                       const comma = i !== call.assigned_unit.length - 1 ? ", " : " ";