+++ client/src/components/admin/AdminSidebar.tsx
@@ -76,7 +76,7 @@ const AdminSidebar: React.FC<Props> = ({ user }) => {
             )} `}
             to="/admin/manage/officers"
           >
-            Officer Management
+            {window.lang.admin.officer_management}
           </Link>
           <Link
             className={`text-decoration-none p-2 rounded admin-link ${isActive(
@@ -84,7 +84,7 @@ const AdminSidebar: React.FC<Props> = ({ user }) => {
             )} `}
             to="/admin/manage/10-codes"
           >
-            10 Codes
+            {window.lang.codes.codes_10}
           </Link>
           <Link
             className={`text-decoration-none p-2 rounded admin-link ${isActive(
@@ -92,7 +92,7 @@ const AdminSidebar: React.FC<Props> = ({ user }) => {
             )} `}
             to="/admin/manage/penal-codes"
           >
-            Penal Codes
+            {window.lang.codes.penal_codes}
           </Link>
           {user?.rank === "owner" ? (
             <Link
@@ -155,6 +155,14 @@ const AdminSidebar: React.FC<Props> = ({ user }) => {
             to="/admin/values/weapons"
           >
             {lang.admin.values.weapons.index}
+          </Link>
+          <Link
+            className={`admin-link text-decoration-none p-2 rounded  ${isActive(
+              "/values/call-types",
+            )}`}
+            to="/admin/values/call-types"
+          >
+            {window.lang.admin.values["call-types"].index}
           </Link>
         </div>
       </div>