+++ client/src/interfaces/Code10.ts
@@ -6,6 +6,7 @@ interface Code10 {
   color: string;
   what_pages: Unit[];
   should_do: string; // should "set_off_duty" or "set_status"
+  position: number;
 }
 
 export default Code10;