+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 1.4.7
+
+- Bug fixes
+- Add missing translation keys
+- Set a position for 10 codes
+- Call types (EG: Citizen call, Traffic stop, ...)
+- Improvements to Docker support, Thanks @bound2 & @henriots!
+
 ## 1.4.6
 
 - New: [Docker support](https://github.com/Dev-CasperTheGhost/snaily-cadv3/wiki/Installation-Guide#using-docker) - for advanced users.